+++ ChangeLog.md
@@ -1,5 +1,14 @@
 # obs2nlm ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Tweaked the builtin preamble that is included in the output file.
+  ([#4](https://github.com/davep/obs2nlm/pull/4))
+- Added an `--instructions` switch so the builtin instructions can be
+  overridden. ([#4](https://github.com/davep/obs2nlm/pull/4))
+
 ## v1.0.0
 
 **Released: 2026-01-08**